more SAF implementation * full support for Directory API (Android Lollipop or later) * best effort to handle any kind errors (missing file, revoked permissions, etc) and recover the download * implemented directory choosing * fix download database version upgrading * misc. cleanup * do not release permission on the old save path (if the user change the download directory) under SAF api

+++ app/src/main/java/us/shandian/giga/io/StoredFileHelper.java
@@ -8,6 +8,7 @@ import android.net.Uri;
 import android.os.Build;
 import android.provider.DocumentsContract;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.provider.DocumentFile;
 
@@ -25,79 +26,52 @@ public class StoredFileHelper implements Serializable {
     private transient DocumentFile docFile;
     private transient DocumentFile docTree;
     private transient File ioFile;
-    private transient ContentResolver contentResolver;
+    private transient Context context;
 
     protected String source;
-    String sourceTree;
+    private String sourceTree;
 
     protected String tag;
 
     private String srcName;
     private String srcType;
 
-    public StoredFileHelper(String filename, String mime, String tag) {
+    public StoredFileHelper(@Nullable Uri parent, String filename, String mime, String tag) {
         this.source = null;// this instance will be "invalid" see invalidate()/isInvalid() methods
 
         this.srcName = filename;
         this.srcType = mime == null ? DEFAULT_MIME : mime;
+        if (parent != null) this.sourceTree = parent.toString();
 
         this.tag = tag;
     }
 
     @TargetApi(Build.VERSION_CODES.LOLLIPOP)
-    StoredFileHelper(DocumentFile tree, ContentResolver contentResolver, String filename, String mime, String tag) throws IOException {
+    StoredFileHelper(@Nullable Context context, DocumentFile tree, String filename, String mime, boolean safe) throws IOException {
         this.docTree = tree;
-        this.contentResolver = contentResolver;
+        this.context = context;
 
-        // this is very slow, because SAF does not allow overwrite
-        DocumentFile res = this.docTree.findFile(filename);
+        DocumentFile res;
 
-        if (res != null && res.exists() && res.isDirectory()) {
-            if (!res.delete())
-                throw new IOException("Directory with the same name found but cannot delete");
-            res = null;
-        }
-
-        if (res == null) {
-            res = this.docTree.createFile(mime == null ? DEFAULT_MIME : mime, filename);
+        if (safe) {
+            // no conflicts (the filename is not in use)
+            res = this.docTree.createFile(mime, filename);
             if (res == null) throw new IOException("Cannot create the file");
+        } else {
+            res = createSAF(context, mime, filename);
         }
 
         this.docFile = res;
-        this.source = res.getUri().toString();
-        this.srcName = getName();
-        this.srcType = getType();
+
+        this.source = docFile.getUri().toString();
+        this.sourceTree = docTree.getUri().toString();
+
+        this.srcName = this.docFile.getName();
+        this.srcType = this.docFile.getType();
     }
 
-    @TargetApi(Build.VERSION_CODES.KITKAT)
-    public StoredFileHelper(Context context, @NonNull Uri path, String tag) throws IOException {
-        this.source = path.toString();
-        this.tag = tag;
-
-        if (path.getScheme() == null || path.getScheme().equalsIgnoreCase("file")) {
-            this.ioFile = new File(URI.create(this.source));
-        } else {
-            DocumentFile file = DocumentFile.fromSingleUri(context, path);
-            if (file == null)
-                throw new UnsupportedOperationException("Cannot get the file via SAF");
-
-            this.contentResolver = context.getContentResolver();
-            this.docFile = file;
-
-            try {
-                this.contentResolver.takePersistableUriPermission(docFile.getUri(), StoredDirectoryHelper.PERMISSION_FLAGS);
-            } catch (Exception e) {
-                throw new IOException(e);
-            }
-        }
-
-        this.srcName = getName();
-        this.srcType = getType();
-    }
-
-    public StoredFileHelper(File location, String filename, String tag) throws IOException {
+    StoredFileHelper(File location, String filename, String mime) throws IOException {
         this.ioFile = new File(location, filename);
-        this.tag = tag;
 
         if (this.ioFile.exists()) {
             if (!this.ioFile.isFile() && !this.ioFile.delete())
@@ -108,22 +82,58 @@ public class StoredFileHelper implements Serializable {
         }
 
         this.source = Uri.fromFile(this.ioFile).toString();
+        this.sourceTree = Uri.fromFile(location).toString();
+
+        this.srcName = ioFile.getName();
+        this.srcType = mime;
+    }
+
+    @TargetApi(Build.VERSION_CODES.KITKAT)
+    public StoredFileHelper(Context context, @Nullable Uri parent, @NonNull Uri path, String tag) throws IOException {
+        this.tag = tag;
+        this.source = path.toString();
+
+        if (path.getScheme() == null || path.getScheme().equalsIgnoreCase(ContentResolver.SCHEME_FILE)) {
+            this.ioFile = new File(URI.create(this.source));
+        } else {
+            DocumentFile file = DocumentFile.fromSingleUri(context, path);
+
+            if (file == null) throw new RuntimeException("SAF not available");
+
+            this.context = context;
+
+            if (file.getName() == null) {
+                this.source = null;
+                return;
+            } else {
+                this.docFile = file;
+                takePermissionSAF();
+            }
+        }
+
+        if (parent != null) {
+            if (!ContentResolver.SCHEME_FILE.equals(parent.getScheme()))
+                this.docTree = DocumentFile.fromTreeUri(context, parent);
+
+            this.sourceTree = parent.toString();
+        }
+
         this.srcName = getName();
         this.srcType = getType();
     }
 
+
     public static StoredFileHelper deserialize(@NonNull StoredFileHelper storage, Context context) throws IOException {
+        Uri treeUri = storage.sourceTree == null ? null : Uri.parse(storage.sourceTree);
+
         if (storage.isInvalid())
-            return new StoredFileHelper(storage.srcName, storage.srcType, storage.tag);
+            return new StoredFileHelper(treeUri, storage.srcName, storage.srcType, storage.tag);
 
-        StoredFileHelper instance = new StoredFileHelper(context, Uri.parse(storage.source), storage.tag);
+        StoredFileHelper instance = new StoredFileHelper(context, treeUri, Uri.parse(storage.source), storage.tag);
 
-        if (storage.sourceTree != null) {
-            instance.docTree = DocumentFile.fromTreeUri(context, Uri.parse(instance.sourceTree));
-
-            if (instance.docTree == null)
-                throw new IOException("Cannot deserialize the tree, ¿revoked permissions?");
-        }
+        // under SAF, if the target document is deleted, conserve the filename and mime
+        if (instance.srcName == null) instance.srcName = storage.srcName;
+        if (instance.srcType == null) instance.srcType = storage.srcType;
 
         return instance;
     }
@@ -143,13 +153,14 @@ public class StoredFileHelper implements Serializable {
         who.startActivityForResult(intent, requestCode);
     }
 
+
     public SharpStream getStream() throws IOException {
         invalid();
 
         if (docFile == null)
             return new FileStream(ioFile);
         else
-            return new FileStreamSAF(contentResolver, docFile.getUri());
+            return new FileStreamSAF(context.getContentResolver(), docFile.getUri());
     }
 
     /**
@@ -173,6 +184,12 @@ public class StoredFileHelper implements Serializable {
         return docFile == null ? Uri.fromFile(ioFile) : docFile.getUri();
     }
 
+    public Uri getParentUri() {
+        invalid();
+
+        return sourceTree == null ? null : Uri.parse(sourceTree);
+    }
+
     public void truncate() throws IOException {
         invalid();
 
@@ -182,17 +199,17 @@ public class StoredFileHelper implements Serializable {
     }
 
     public boolean delete() {
-        invalid();
-
+        if (source == null) return true;
         if (docFile == null) return ioFile.delete();
+
 
         boolean res = docFile.delete();
 
         try {
             int flags = Intent.FLAG_GRANT_READ_URI_PERMISSION | Intent.FLAG_GRANT_WRITE_URI_PERMISSION;
-            contentResolver.releasePersistableUriPermission(docFile.getUri(), flags);
+            context.getContentResolver().releasePersistableUriPermission(docFile.getUri(), flags);
         } catch (Exception ex) {
-            // ¿what happen?
+            // nothing to do
         }
 
         return res;
@@ -209,18 +226,22 @@ public class StoredFileHelper implements Serializable {
         return docFile == null ? ioFile.canWrite() : docFile.canWrite();
     }
 
-    public File getIOFile() {
-        return ioFile;
-    }
-
     public String getName() {
-        if (source == null) return srcName;
-        return docFile == null ? ioFile.getName() : docFile.getName();
+        if (source == null)
+            return srcName;
+        else if (docFile == null)
+            return ioFile.getName();
+
+        String name = docFile.getName();
+        return name == null ? srcName : name;
     }
 
     public String getType() {
-        if (source == null) return srcType;
-        return docFile == null ? DEFAULT_MIME : docFile.getType();// not obligatory for Java IO
+        if (source == null || docFile == null)
+            return srcType;
+
+        String type = docFile.getType();
+        return type == null ? srcType : type;
     }
 
     public String getTag() {
@@ -231,29 +252,41 @@ public class StoredFileHelper implements Serializable {
         if (source == null) return false;
 
         boolean exists = docFile == null ? ioFile.exists() : docFile.exists();
-        boolean asFile = docFile == null ? ioFile.isFile() : docFile.isFile();// ¿docFile.isVirtual() means is no-physical?
+        boolean isFile = docFile == null ? ioFile.isFile() : docFile.isFile();// ¿docFile.isVirtual() means is no-physical?
 
-        return exists && asFile;
+        return exists && isFile;
     }
 
     public boolean create() {
         invalid();
+        boolean result;
 
         if (docFile == null) {
             try {
-                return ioFile.createNewFile();
+                result = ioFile.createNewFile();
+            } catch (IOException e) {
+                return false;
+            }
+        } else if (docTree == null) {
+            result = false;
+        } else {
+            if (!docTree.canRead() || !docTree.canWrite()) return false;
+            try {
+                docFile = createSAF(context, srcType, srcName);
+                if (docFile == null || docFile.getName() == null) return false;
+                result = true;
             } catch (IOException e) {
                 return false;
             }
         }
 
-        if (docTree == null || docFile.getName() == null) return false;
+        if (result) {
+            source = (docFile == null ? Uri.fromFile(ioFile) : docFile.getUri()).toString();
+            srcName = getName();
+            srcType = getType();
+        }
 
-        DocumentFile res = docTree.createFile(docFile.getName(), docFile.getType() == null ? DEFAULT_MIME : docFile.getType());
-        if (res == null) return false;
-
-        docFile = res;
-        return true;
+        return result;
     }
 
     public void invalidate() {
@@ -264,20 +297,25 @@ public class StoredFileHelper implements Serializable {
 
         source = null;
 
-        sourceTree = null;
         docTree = null;
         docFile = null;
         ioFile = null;
-        contentResolver = null;
-    }
-
-    private void invalid() {
-        if (source == null)
-            throw new IllegalStateException("In invalid state");
+        context = null;
     }
 
     public boolean equals(StoredFileHelper storage) {
-        if (this.isInvalid() != storage.isInvalid()) return false;
+        if (this == storage) return true;
+
+        // note: do not compare tags, files can have the same parent folder
+        //if (stringMismatch(this.tag, storage.tag)) return false;
+
+        if (stringMismatch(getLowerCase(this.sourceTree), getLowerCase(this.sourceTree)))
+            return false;
+
+        if (this.isInvalid() || storage.isInvalid()) {
+            return this.srcName.equalsIgnoreCase(storage.srcName) && this.srcType.equalsIgnoreCase(storage.srcType);
+        }
+
         if (this.isDirect() != storage.isDirect()) return false;
 
         if (this.isDirect())
@@ -297,5 +335,47 @@ public class StoredFileHelper implements Serializable {
             return "[Invalid state] name=" + srcName + "  type=" + srcType + "  tag=" + tag;
         else
             return "sourceFile=" + source + "  treeSource=" + (sourceTree == null ? "" : sourceTree) + "  tag=" + tag;
+    }
+
+
+    private void invalid() {
+        if (source == null)
+            throw new IllegalStateException("In invalid state");
+    }
+
+    private void takePermissionSAF() throws IOException {
+        try {
+            context.getContentResolver().takePersistableUriPermission(docFile.getUri(), StoredDirectoryHelper.PERMISSION_FLAGS);
+        } catch (Exception e) {
+            if (docFile.getName() == null) throw new IOException(e);
+        }
+    }
+
+    private DocumentFile createSAF(@Nullable Context context, String mime, String filename) throws IOException {
+        DocumentFile res = StoredDirectoryHelper.findFileSAFHelper(context, docTree, filename);
+
+        if (res != null && res.exists() && res.isDirectory()) {
+            if (!res.delete())
+                throw new IOException("Directory with the same name found but cannot delete");
+            res = null;
+        }
+
+        if (res == null) {
+            res = this.docTree.createFile(srcType == null ? DEFAULT_MIME : mime, filename);
+            if (res == null) throw new IOException("Cannot create the file");
+        }
+
+        return res;
+    }
+
+    private String getLowerCase(String str) {
+        return str == null ? null : str.toLowerCase();
+    }
+
+    private boolean stringMismatch(String str1, String str2) {
+        if (str1 == null && str2 == null) return false;
+        if ((str1 == null) != (str2 == null)) return true;
+
+        return !str1.equals(str2);
     }
 }
